feat(benchmarks): package benchmark suite into misda.benchmarks subpackage with optional extra
