+++ pyproject.toml
@@ -25,11 +25,14 @@ dependencies = [
 ]
 requires-python = ">=3.8"
 
+[project.optional-dependencies]
+benchmarks = []
+
 [project.urls]
 Homepage = "https://github.com/monacofj/misda"
 
 [tool.setuptools]
-packages = ["misda"]
+packages = ["misda", "misda.benchmarks"]
 
 [tool.pytest.ini_options]
 markers = [
